LayoutTests/http/tests/security/no-javascript-refresh-percent-escaped-expected.txt

PASS: This is the content that appears in place of a failed refresh.
